requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
groq==0.9.0
requests==2.31.0
pdfkit==1.0.0